Introduce a DBC class for contracts (#12)
This patch introduces a separate class to encapsulate the contract logic
as it is quite hard to read the code in a reversed logic
(if-not-postcondition-then-throw).

Additionally, as not-null checks obscured the code coverage, we
introduce a not-null assertion function. To that end, we had to use
`NotNullAttribute` which is available only in NET Standard 2.1. This
forced us thus to bump to NET Standard 2.1 from 2.0.