Add support for netstandard 2.0
In this patch, we dispense of the features from netstandard 2.1 (such as
`ToHashSet`) so that we can support netstandard 2.0 as well.

Notably, this support is necessary so that we can use the library
together with aasx-package-explorer.

Unfortunately, we could not use NotNull annotation anymore, so
assertions could not be ignored in the code coverage finally
resulting in lower coverage.