Bump System.IO.Packaging from 4.* to 9.* (#46)
We drop support for net5.0 and net6.0 and add support for net8.0 and
net9.0 so that we can use `System.IO.Packaging` 9.* due to
vulnerabilities.

The change to net8.0 and above forced us to remove most of the
style-checking tools since they were all still operating on net6.0.

We also bump `Microsoft.NET.Test.Sdk` from 17.1.0 to 17.13.0 to fix
(other) vulnerabilities.

The original pull request was #44, but it got stale. This change
supplements it.

Fixes #43.